--- omegaconf/resolvers/oc/_decode.py
@@ -1,25 +1,0 @@
-from typing import Any, Optional
-
-from omegaconf import Container
-from omegaconf._utils import _get_value
-from omegaconf.grammar_parser import parse
-
-
-def decode(expr: Optional[str], _parent_: Container) -> Any:
-    """
-    Parse and evaluate `expr` according to the `singleElement` rule of the grammar.
-
-    If `expr` is `None`, then return `None`.
-    """
-    if expr is None:
-        return None
-
-    if not isinstance(expr, str):
-        raise TypeError(
-            f"`oc.decode` can only take strings or None as input, "
-            f"but `{expr}` is of type {type(expr).__name__}"
-        )
-
-    parse_tree = parse(expr, parser_rule="singleElement", lexer_mode="VALUE_MODE")
-    val = _parent_.resolve_parse_tree(parse_tree)
-    return _get_value(val)
