--- omegaconf/resolvers/oc/dict/__init__.py
@@ -1,7 +1,0 @@
-from ._keys import dict_keys
-from ._values import dict_values
-
-__all__ = [
-    "dict_keys",
-    "dict_values",
-]
