+++ setup.py
@@ -56,7 +56,6 @@ with open("README.md", "r") as fh:
             "omegaconf.grammar.gen",
             "omegaconf.resolvers",
             "omegaconf.resolvers.oc",
-            "omegaconf.resolvers.oc.dict",
             "pydevd_plugins",
             "pydevd_plugins.extensions",
         ],
